+++ relayer/chains/icon/provider.go
@@ -25,7 +25,8 @@ type Config struct {
 	NID            string                         `json:"nid" yaml:"nid"`
 	StepMin        int64                          `json:"step-min" yaml:"step-min"`
 	StepLimit      int64                          `json:"step-limit" yaml:"step-limit"`
-	StepAdjustment int                            `json:"step-adjustment" yaml:"step-adjustment"`
+	StepAdjustment int64                          `json:"step-adjustment" yaml:"step-adjustment"`
+	Disabled       bool                           `json:"disabled" yaml:"disabled"`
 	HomeDir        string                         `json:"-" yaml:"-"`
 }
 
@@ -67,7 +68,7 @@ func (c *Config) Validate() error {
 
 func (c *Config) sanitize() error {
 	if c.StepAdjustment == 0 {
-		c.StepAdjustment = 10
+		c.StepAdjustment = 50
 	}
 	return nil
 }
@@ -78,6 +79,11 @@ func (p *Config) SetWallet(addr string) {
 
 func (p *Config) GetWallet() string {
 	return p.Address
+}
+
+// Enabled returns true if the chain is enabled
+func (c *Config) Enabled() bool {
+	return !c.Disabled
 }
 
 type Provider struct {