+++ relayer/chains/wasm/config.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"crypto/tls"
 	"fmt"
+	"os"
 	"path/filepath"
 	"time"
 
@@ -19,7 +20,7 @@ import (
 	"google.golang.org/grpc/credentials"
 )
 
-type ProviderConfig struct {
+type Config struct {
 	RpcUrl                 string                          `json:"rpc-url" yaml:"rpc-url"`
 	GrpcUrl                string                          `json:"grpc-url" yaml:"grpc-url"`
 	ChainID                string                          `json:"chain-id" yaml:"chain-id"`
@@ -41,10 +42,12 @@ type ProviderConfig struct {
 	Simulate               bool                            `json:"simulate" yaml:"simulate"`
 	StartHeight            uint64                          `json:"start-height" yaml:"start-height"`
 	FinalityBlock          uint64                          `json:"finality-block" yaml:"finality-block"`
+	Disabled               bool                            `json:"disabled" yaml:"disabled"`
+	ExtraCodec             string                          `json:"extra-codecs" yaml:"extra-codecs"`
 	ChainName              string                          `json:"-" yaml:"-"`
 }
 
-func (pc *ProviderConfig) NewProvider(ctx context.Context, log *zap.Logger, homePath string, _ bool, chainName string) (provider.ChainProvider, error) {
+func (pc *Config) NewProvider(ctx context.Context, log *zap.Logger, homePath string, _ bool, chainName string) (provider.ChainProvider, error) {
 	pc.HomeDir = homePath
 	pc.ChainName = chainName
 
@@ -79,15 +82,15 @@ func (pc *ProviderConfig) NewProvider(ctx context.Context, log *zap.Logger, home
 	}, nil
 }
 
-func (pc *ProviderConfig) SetWallet(addr string) {
+func (pc *Config) SetWallet(addr string) {
 	pc.Address = addr
 }
 
-func (pc *ProviderConfig) GetWallet() string {
+func (pc *Config) GetWallet() string {
 	return pc.Address
 }
 
-func (pc *ProviderConfig) Validate() error {
+func (pc *Config) Validate() error {
 	if pc.ChainName == "" {
 		return fmt.Errorf("chain-name cannot be empty")
 	}
@@ -98,18 +101,22 @@ func (pc *ProviderConfig) Validate() error {
 	return nil
 }
 
-func (pc *ProviderConfig) sanitize() (*ProviderConfig, error) {
+func (c *Config) Enabled() bool {
+	return !c.Disabled
+}
+
+func (pc *Config) sanitize() (*Config, error) {
 	return pc, nil
 }
 
-func (c *ProviderConfig) newClientContext(ctx context.Context) (*sdkClient.Context, error) {
-	codec := GetCodecConfig(c)
+func (c *Config) newClientContext(ctx context.Context) (sdkClient.Context, error) {
+	codec := c.MakeCodec(moduleBasics, c.ExtraCodec)
 
 	keyRing, err := keyring.New(
 		c.ChainName,
 		c.KeyringBackend,
 		c.KeyringDir,
-		nil,
+		os.Stdin,
 		codec.Codec,
 		func(options *keyring.Options) {
 			options.SupportedAlgos = types.SupportedAlgorithms
@@ -117,20 +124,21 @@ func (c *ProviderConfig) newClientContext(ctx context.Context) (*sdkClient.Conte
 		},
 	)
 	if err != nil {
-		return nil, err
+		return sdkClient.Context{}, err
 	}
 
 	cometRPCClient, err := http.New(c.RpcUrl, "/websocket")
 	if err != nil {
-		return nil, err
+		return sdkClient.Context{}, err
 	}
 
-	grpcClient, err := grpc.DialContext(ctx, c.GrpcUrl, grpc.WithTransportCredentials(credentials.NewTLS(&tls.Config{InsecureSkipVerify: true})))
+	grpcClient, err := grpc.NewClient(c.GrpcUrl, grpc.WithTransportCredentials(credentials.NewTLS(&tls.Config{})))
 	if err != nil {
-		return nil, err
+		return sdkClient.Context{}, err
 	}
 
-	return &sdkClient.Context{
+	return sdkClient.Context{
+		CmdContext:        ctx,
 		ChainID:           c.ChainID,
 		Client:            cometRPCClient,
 		NodeURI:           c.RpcUrl,